Migrate project to Python 3 and stabilize test suite

+++ fit/fit/PrimitiveFixture.py
@@ -15,7 +15,7 @@ class PrimitiveFixture(Fixture):
     ## format converters ########################
 
     def parseLong(self,cell):
-        return long(cell.text())
+        return int(cell.text())
 
     ## answer comparisons ######################
